+++ src/steam_client.h
@@ -42,6 +42,23 @@ class SteamClient {
     virtual void OnGameLobbyJoinRequested(uint64 SteamIdLobby, uint64 SteamIdUser) = 0;
     virtual void OnGameRichPresenceJoinRequested(uint64 steamIDFriend, std::string rgchConnect) = 0;
     virtual void OnNewUrlLaunchParameters() = 0;
+    virtual void OnFloatingGamepadTextInputDismissed() = 0;
+
+    virtual void OnLobbyMatchList(uint32 LobbiesMatching) = 0;
+    virtual void OnP2PSessionRequest(uint64 steamIDRemote) = 0;
+    virtual void OnP2PSessionConnectFail(uint64 steamIDRemote,
+                                         uint8 eP2PSessionError) = 0;
+    virtual void OnLobbyChatMsg(uint64 steamIDLobby, uint64 steamIDUser,
+                                uint8 chatEntryType, uint32 chatID) = 0;
+    virtual void OnLobbyChatUpdate(uint64 SteamIDLobby,
+                                   uint64 SteamIDUserChanged,
+                                   uint64 SteamIDMakingChange,
+                                   uint32 ChatMemberStateChange) = 0;
+
+    virtual void OnValidateAuthTicketResponse(CSteamID m_SteamID,
+                                   EAuthSessionResponse m_eAuthSessionResponse,
+                                   CSteamID m_OwnerSteamID) = 0;
+
     virtual ~Observer() {}
   };
 
@@ -90,9 +107,29 @@ class SteamClient {
   STEAM_CALLBACK(SteamClient, OnLobbyDataUpdate, LobbyDataUpdate_t, OnLobbyDataUpdate_);
   STEAM_CALLBACK(SteamClient, OnLobbyEnter, LobbyEnter_t, OnLobbyEnter_);
   STEAM_CALLBACK(SteamClient, OnLobbyInvite, LobbyInvite_t, OnLobbyInvite_);
-  STEAM_CALLBACK(SteamClient, OnGameLobbyJoinRequested, GameLobbyJoinRequested_t, OnGameLobbyJoinRequested_);
-  STEAM_CALLBACK(SteamClient, OnGameRichPresenceJoinRequested, GameRichPresenceJoinRequested_t, OnGameRichPresenceJoinRequested_);
-  STEAM_CALLBACK(SteamClient, OnNewUrlLaunchParameters, NewUrlLaunchParameters_t, OnNewUrlLaunchParameters_);
+  STEAM_CALLBACK(SteamClient, OnGameLobbyJoinRequested,
+                 GameLobbyJoinRequested_t, OnGameLobbyJoinRequested_);
+  STEAM_CALLBACK(SteamClient, OnGameRichPresenceJoinRequested,
+                 GameRichPresenceJoinRequested_t,
+                 OnGameRichPresenceJoinRequested_);
+  STEAM_CALLBACK(SteamClient, OnNewUrlLaunchParameters,
+                 NewUrlLaunchParameters_t, OnNewUrlLaunchParameters_);
+  STEAM_CALLBACK(SteamClient, OnFloatingGamepadTextInputDismissed,
+                 FloatingGamepadTextInputDismissed_t,
+                 OnFloatingGamepadTextInputDismissed_);
+
+  STEAM_CALLBACK(SteamClient, OnLobbyMatchList, LobbyMatchList_t,
+                 OnLobbyMatchList_);
+  STEAM_CALLBACK(SteamClient, OnP2PSessionRequest, P2PSessionRequest_t,
+                 OnP2PSessionRequest_);
+  STEAM_CALLBACK(SteamClient, OnP2PSessionConnectFail, P2PSessionConnectFail_t,
+                 OnP2PSessionConnectFail_);
+  STEAM_CALLBACK(SteamClient, OnLobbyChatMsg, LobbyChatMsg_t, OnLobbyChatMsg_);
+  STEAM_CALLBACK(SteamClient, OnLobbyChatUpdate, LobbyChatUpdate_t,
+                 OnLobbyChatUpdate_);
+
+
+  STEAM_CALLBACK(SteamClient, OnValidateAuthTicketResponse, ValidateAuthTicketResponse_t, OnValidateAuthTicketResponse_);
 };
 
 }  // namespace greenworks