+++ src/steam_event.h
@@ -31,18 +31,37 @@ class SteamEvent : public greenworks::SteamClient::Observer {
   void OnGameConnectedFriendChatMessage(uint64 raw_steam_id,
                                         int message_id) override;
   void OnDLCInstalled(AppId_t dlc_app_id) override;
-  void OnMicroTxnAuthorizationResponse(uint32 AppID,
-                                       uint64 OrderID,
+  void OnMicroTxnAuthorizationResponse(uint32 AppID, uint64 OrderID,
                                        bool Autorized) override;
-  void OnLobbyCreated(int status_code, uint64 SteamIdLobby);
-  void OnLobbyDataUpdate(uint64 SteamIdLobby, uint64 SteamIdMember, bool Success);
-  void OnLobbyEnter(uint64 SteamIdLobby, int ChatPermissions, bool Locked, int ChatRoomEnterResponse);
-  void OnLobbyInvite(uint64 SteamIdUser, uint64 SteamIdLobby, uint64 GameId);
-  void OnGameLobbyJoinRequested(uint64 SteamIdLobby, uint64 SteamIdUser);
-  void OnGameRichPresenceJoinRequested(uint64 steamIDFriend, std::string rgchConnect);
-  void OnNewUrlLaunchParameters();
+  void OnLobbyCreated(int status_code, uint64 SteamIdLobby) override;
+  void OnLobbyDataUpdate(uint64 SteamIdLobby, uint64 SteamIdMember,
+                         bool Success) override;
+  void OnLobbyEnter(uint64 SteamIdLobby, int ChatPermissions, bool Locked,
+                    int ChatRoomEnterResponse) override;
+  void OnLobbyInvite(uint64 SteamIdUser, uint64 SteamIdLobby,
+                     uint64 GameId) override;
+  void OnGameLobbyJoinRequested(uint64 SteamIdLobby,
+                                uint64 SteamIdUser) override;
+  void OnGameRichPresenceJoinRequested(uint64 steamIDFriend,
+                                       std::string rgchConnect) override;
+  void OnNewUrlLaunchParameters() override;
+  void OnFloatingGamepadTextInputDismissed() override;
 
- private:
+  void OnLobbyMatchList(uint32 LobbiesMatching) override;
+  void OnP2PSessionRequest(uint64 steamIDRemote) override;
+  void OnP2PSessionConnectFail(uint64 steamIDRemote,
+                               uint8 eP2PSessionError) override;
+  void OnLobbyChatMsg(uint64 steamIDLobby, uint64 steamIDUser,
+                      uint8 chatEntryType, uint32 chatID) override;
+  void OnLobbyChatUpdate(uint64 SteamIDLobby, uint64 SteamIDUserChanged,
+                         uint64 SteamIDMakingChange,
+                         uint32 ChatMemberStateChange) override;
+
+  void OnValidateAuthTicketResponse(CSteamID m_SteamID,
+                         EAuthSessionResponse m_eAuthSessionResponse,
+                         CSteamID m_OwnerSteamID) override;
+
+private:
   const Nan::Persistent<v8::Object>& persistent_steam_events_;
 };