+++ src/steam_id.cc
@@ -45,105 +45,105 @@ v8::Local<v8::Object> SteamID::Create(CSteamID steam_id) {
 }
 
 NAN_METHOD(SteamID::IsAnonymous) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(obj->steam_id_.BAnonAccount());
 }
 
 NAN_METHOD(SteamID::IsAnonymousGameServer) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(obj->steam_id_.BAnonGameServerAccount());
 }
 
 NAN_METHOD(SteamID::IsAnonymousGameServerLogin) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.BBlankAnonAccount()));
 }
 
 NAN_METHOD(SteamID::IsAnonymousUser) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.BAnonUserAccount()));
 }
 
 NAN_METHOD(SteamID::IsChatAccount) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.BChatAccount()));
 }
 
 NAN_METHOD(SteamID::IsClanAccount) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.BClanAccount()));
 }
 
 NAN_METHOD(SteamID::IsConsoleUserAccount) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.BConsoleUserAccount()));
 }
 
 NAN_METHOD(SteamID::IsContentServerAccount) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.BContentServerAccount()));
 }
 
 NAN_METHOD(SteamID::IsGameServerAccount) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.BGameServerAccount()));
 }
 
 NAN_METHOD(SteamID::IsIndividualAccount) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.BIndividualAccount()));
 }
 
 NAN_METHOD(SteamID::IsPersistentGameServerAccount) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(
       Nan::New(obj->steam_id_.BPersistentGameServerAccount()));
 }
 
 NAN_METHOD(SteamID::IsLobby) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.IsLobby()));
 }
 
 NAN_METHOD(SteamID::GetAccountID) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(
       Nan::New<v8::Integer>(obj->steam_id_.GetAccountID()));
 }
 
 NAN_METHOD(SteamID::GetRawSteamID) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(
       Nan::New(utils::uint64ToString(obj->steam_id_.ConvertToUint64()))
           .ToLocalChecked());
 }
 
 NAN_METHOD(SteamID::GetAccountType) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.GetEAccountType()));
 }
 
 NAN_METHOD(SteamID::IsValid) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(Nan::New(obj->steam_id_.IsValid()));
 }
 
 NAN_METHOD(SteamID::GetStaticAccountKey) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(
       Nan::New(utils::uint64ToString(
           obj->steam_id_.GetStaticAccountKey())).ToLocalChecked());
 }
 
 NAN_METHOD(SteamID::GetPersonaName) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(
       Nan::New(SteamFriends()->GetFriendPersonaName(obj->steam_id_))
           .ToLocalChecked());
 }
 
 NAN_METHOD(SteamID::GetNickname) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   const char* nick_name = SteamFriends()->GetPlayerNickname(obj->steam_id_);
   if (nick_name) {
     info.GetReturnValue().Set(Nan::New(nick_name).ToLocalChecked());
@@ -153,13 +153,13 @@ NAN_METHOD(SteamID::GetNickname) {
 }
 
 NAN_METHOD(SteamID::GetRelationship) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(
       Nan::New(SteamFriends()->GetFriendRelationship(obj->steam_id_)));
 }
 
 NAN_METHOD(SteamID::GetSteamLevel) {
-  auto* obj = ObjectWrap::Unwrap<SteamID>(info.Holder());
+  auto* obj = ObjectWrap::Unwrap<SteamID>(info.This());
   info.GetReturnValue().Set(
       Nan::New(SteamFriends()->GetFriendSteamLevel(obj->steam_id_)));
 }